性能测试新架构，避免平台自身性能问题

+++ testcenterservice/src/main/java/com/zoctan/api/dto/FunctionCaseSandF.java
@@ -22,14 +22,14 @@ public class FunctionCaseSandF {
     }
 
 
-    public int getValue() {
+    public long getValue() {
         return value;
     }
 
-    public void setValue(int value) {
+    public void setValue(long value) {
         this.value = value;
     }
 
-    private int  value;
+    private long  value;
 
 }
